requirements.txt
pandas
numpy
matplotlib
torch
torchvision
Pillow
scikit-image
